requirements.txt
pandas==2.0.3
python-miio==0.5.12
requests==2.31.0
